Allow creation of content elements
Content elements could not be created, since they needed campaigns.
In order to retrieve campaigns, the API needs to be created.
The API needs settings from current site, which is not available in that
context.

Therefore the factory is extended to also fetch site and their settings
based on page uid. The page uid is available in the context.

Since we now interact with guzzle, we add the dependency.
Wouldn't be necessary if SDK would have PSR compatible type hints.

Also tests were added for code, to keep code coverage.

Fixes: #5